Added some old scripts
I got a lot of questions in my tutorial video, whether if I could share the code that I was using inside Unity showcase project. Only problem was that I delete project (because it was a temporary project). However, luckily I was smart enough and exported the code into a unity package, before deleting the project. By using unitypackage_extractor: https://github.com/Cobertos/unitypackage_extractor, I was able to extract the content out and add it to this repo.